Check the number of translated strings after loading the resource
For EmuTOS, when new strings to be translated are added to the
resource, the messages.pot is not necessarily updated, so the
previous behaviour of reporting the number of untranslated strings
from the messages.pot was not convenient. Check the number now
after loading the resource, using the same exceptions as erd.